Handle different exceptions

+++ TABP.Application/Commands/Amenities/DeleteAmenity/DeleteAmenityCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using MediatR;
 using TABP.DAL.Interfaces;
 using TABP.DAL.Interfaces.Repositories;
+using TABP.Domain.Exceptions;
 
 namespace TABP.Application.Commands.Amenities.DeleteAmenity;
 
@@ -17,6 +18,11 @@ public class DeleteAmenityCommandHandler : IRequestHandler<DeleteAmenityCommand>
 
     public async Task Handle(DeleteAmenityCommand request, CancellationToken cancellationToken)
     {
+        if (!await _amenityRepository.ExistsAsync(a => a.Id == request.Id))
+        {
+            throw new NotFoundException($"Amenity with id {request.Id} not found");
+        }
+
         await _amenityRepository.DeleteAsync(request.Id);
 
         await _unitOfWork.SaveChangesAsync();
