+++ TABP.Application/Commands/Hotels/DeleteHotel/DeleteHotelCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using MediatR;
 using TABP.DAL.Interfaces;
 using TABP.DAL.Interfaces.Repositories;
+using TABP.Domain.Exceptions;
 
 namespace TABP.Application.Commands.Hotels.DeleteHotel;
 
@@ -17,6 +18,11 @@ public class DeleteHotelCommandHandler : IRequestHandler<DeleteHotelCommand>
 
     public async Task Handle(DeleteHotelCommand request, CancellationToken cancellationToken)
     {
+        if (!await _hotelRepository.ExistsAsync(h => h.Id == request.Id))
+        {
+            throw new NotFoundException($"Hotel with id {request.Id} wasn't found");
+        }
+
         await _hotelRepository.DeleteAsync(request.Id);
 
         await _unitOfWork.SaveChangesAsync();
