+++ TABP.Application/Commands/RoomAmenities/DeleteRoomAmenity/DeleteRoomAmenityCommandHandler.cs
@@ -2,6 +2,7 @@
 using MediatR;
 using TABP.DAL.Interfaces;
 using TABP.DAL.Interfaces.Repositories;
+using TABP.Domain.Exceptions;
 
 namespace TABP.Application.Commands.RoomAmenities.DeleteRoomAmenity;
 
@@ -19,6 +20,10 @@ public class DeleteRoomAmenityCommandHandler : IRequestHandler<DeleteRoomAmenity
 
     public async Task Handle(DeleteRoomAmenityCommand request, CancellationToken cancellationToken)
     {
+        if (!await _roomAmenityRepository.ExistsAsync(ra => ra.Id == request.Id
+                                                            && ra.RoomId == request.RoomId))
+            throw new NotFoundException($"Room amenity with id {request.Id} wasn't found");
+
         await _roomAmenityRepository.DeleteAsync(request.Id);
 
         await _unitOfWork.SaveChangesAsync();
